Add files via upload
This is the initial version, we have to start the scraping process using our system manually. Basically the raw version for easier understanding
The next version "AmazonUrlScraper2.0" will automate the process using the function but it requires the manual task to start the kernel only once after that it keeps on running until our system is shut down.